seven segment now works in indefinite loop
flicks between all digits and displays the each value on the seven segment. Problem is atm the code cannot break from this while in normal operating mode because it can't leave the loop in order to continue displaying multiple values. Easiest fix is just to display one value at a time